refactor: Service가 추상적인 DAO에 의존하도록 변경
