+++ src/test/java/racingcar/dao/JdbcTemplateRacingGameDaoTest.java
@@ -1,10 +1,12 @@
 package racingcar.dao;
 
+import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.jdbc.core.JdbcTemplate;
 import racingcar.dto.CarDto;
 
@@ -13,16 +15,19 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
-class RacingGameDaoTest {
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class JdbcTemplateRacingGameDaoTest {
 
+    @LocalServerPort
+    private int port;
     @Autowired
-    private RacingGameDao racingGameDao;
+    private JdbcTemplateRacingGameDao jdbcTemplateRacingGameDao;
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
     @BeforeEach
     void setUp() {
+        RestAssured.port = port;
         jdbcTemplate.execute("DROP TABLE PLAYER_RESULT IF EXISTS");
         jdbcTemplate.execute("DROP TABLE GAME_RESULT IF EXISTS");
 
@@ -43,7 +48,7 @@ class RacingGameDaoTest {
     @DisplayName("게임 결과가 데이터베이스에 저장될 수 있다")
     void shouldSaveGameResultWhenRequest() {
         String expectedWinners = "브리,브라운";
-        racingGameDao.saveGameResult(expectedWinners, 3);
+        jdbcTemplateRacingGameDao.saveGameResult(expectedWinners, 3);
 
         String actualWinners = jdbcTemplate.queryForObject("SELECT winners FROM GAME_RESULT", String.class);
         int trialCount = jdbcTemplate.queryForObject("SELECT trial_count FROM GAME_RESULT", Integer.class);
@@ -56,7 +61,7 @@ class RacingGameDaoTest {
     void shouldSaveEachPlayerResultWhenRequest() {
         // GAME_RESULT 를 저장한다
         String expectedWinners = "브리,브라운";
-        Number gameResultKey = racingGameDao.saveGameResult(expectedWinners, 3);
+        Number gameResultKey = jdbcTemplateRacingGameDao.saveGameResult(expectedWinners, 3);
 
         List<CarDto> carDtos = List.of(
                 new CarDto("브리", "2"),
@@ -64,7 +69,7 @@ class RacingGameDaoTest {
                 new CarDto("브라운", "2")
         );
         // 저장한 GAME_RESULT 을 참조하는 PLAYER_RESULT 를 저장한다
-        racingGameDao.savePlayerResults(carDtos, gameResultKey);
+        jdbcTemplateRacingGameDao.savePlayerResults(carDtos, gameResultKey);
 
         int positionOfBri = jdbcTemplate.queryForObject("SELECT position FROM PLAYER_RESULT WHERE name = '브리'", Integer.class);
         int positionOfTomi = jdbcTemplate.queryForObject("SELECT position FROM PLAYER_RESULT WHERE name = '토미'", Integer.class);
