+++ src/test/java/racingcar/service/RacingGameServiceTest.java
@@ -13,7 +13,7 @@ class RacingGameServiceTest {
     @Test
     @DisplayName("플레이어가 한명일 때 GameResultDto가 잘 반환된다")
     void shouldReturnCorrectlyWhenInputOnePlayer() {
-        RacingGameService racingGameService = new RacingGameService(null);
+        RacingGameService racingGameService = new RacingGameService(new DummyRacingGameDao());
         RacingGameRequestDto racingGameRequestDto = new RacingGameRequestDto(
                 "브리",
                 3
@@ -28,7 +28,7 @@ class RacingGameServiceTest {
     @Test
     @DisplayName("플레이어가 여려명일 때 GameResultDto가 잘 반환된다")
     void shouldReturnCorrectlyWhenInputManyPlayer() {
-        RacingGameService racingGameService = new RacingGameService(null);
+        RacingGameService racingGameService = new RacingGameService(new DummyRacingGameDao());
         RacingGameRequestDto racingGameRequestDto = new RacingGameRequestDto(
                 "브리,토미,브라운",
                 3
